refactor: re-organise dependency injection modules such as fake repo's, codes, and repo hilt module
