+++ app/src/test/java/com/example/librewards/dependency_injection/TestRepositoryModule.java
@@ -1,13 +1,13 @@
-package com.example.librewards;
+package com.example.librewards.dependency_injection;
 
 import com.example.librewards.data.repositories.RewardsRepository;
 import com.example.librewards.data.repositories.StartCodesRepository;
 import com.example.librewards.data.repositories.StopCodesRepository;
 import com.example.librewards.data.repositories.UserRepository;
-import com.example.librewards.repositories.RewardsRepositoryFake;
-import com.example.librewards.repositories.StartCodesRepositoryFake;
-import com.example.librewards.repositories.StopCodesRepositoryFake;
-import com.example.librewards.repositories.UserRepositoryFake;
+import com.example.librewards.dependency_injection.repositories.RewardsRepositoryFake;
+import com.example.librewards.dependency_injection.repositories.StartCodesRepositoryFake;
+import com.example.librewards.dependency_injection.repositories.StopCodesRepositoryFake;
+import com.example.librewards.dependency_injection.repositories.UserRepositoryFake;
 
 import javax.inject.Singleton;
